Corrected prototype of mpfr_can_round_raw.
[From SVN r121 (trunk)]

+++ mpfr.h
@@ -85,7 +85,8 @@ int mpfr_round_raw2 _PROTO ((mp_limb_t *, unsigned long, char, char,
 void mpfr_round _PROTO ((mpfr_ptr, char, unsigned long)); 
 int mpfr_can_round _PROTO ((mpfr_ptr, unsigned long, unsigned char, 
 			    unsigned char, unsigned long));
-int mpfr_can_round_raw _PROTO ((mp_limb_t *, unsigned char, unsigned long, 
+int mpfr_can_round_raw _PROTO ((mp_limb_t *, unsigned long, char, 
+				unsigned long, 
 				unsigned char, unsigned char, unsigned long));
 void mpfr_set_d _PROTO ((mpfr_ptr, double, unsigned char)); 
 int mpfr_set_z _PROTO ((mpfr_ptr, mpz_srcptr, unsigned char));
